// backend/src/db/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  PARENT: 'PARENT',
  CHILD: 'CHILD'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TransactionType = {
  BUY: 'BUY',
  SELL: 'SELL'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const NotificationType = {
  INFO: 'INFO',
  WARNING: 'WARNING',
  ACHIEVEMENT: 'ACHIEVEMENT'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const Sentiment = {
  POSITIVE: 'POSITIVE',
  NEGATIVE: 'NEGATIVE',
  NEUTRAL: 'NEUTRAL'
} as const

export type Sentiment = (typeof Sentiment)[keyof typeof Sentiment]